pyproject.toml
[project]
name = "akinator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.2",
    "google-genai>=1.62.0",
]
